refactor: extract email form logic into custom hook
- Moved email form state and submission logic into `useEmailForm` hook.
- Addresses PR feedback regarding separation of complexity from UI/Layout components.
- Adheres to AGENTS.md Rule 6 (Business logic in hooks).
